+++ Sources/DKImageDataManager/Model/DKAssetGroup.swift
@@ -9,6 +9,7 @@
 import Photos
 
 /// A representation of a Photos asset grouping, such as a moment, user-created album, or smart album.
+@MainActor
 public class DKAssetGroup: NSObject {
     public let groupId: String