+++ Sources/Extensions/DKImageExtensionCamera.swift
@@ -46,9 +46,11 @@ open class DKImageExtensionCamera: DKImageBaseExtension {
         }
         
         func setup() {
-            camera.cameraOverlayView = nil
+            DispatchQueue.main.async {
+                camera.cameraOverlayView = nil
+            }
         }
-        
+
         DKCamera.checkCameraPermission { granted in
             granted ? setup() : cameraDenied()
         }